Continuing work on refactoring.
Base elements needed for simulation present (lightsourced added. Yet, it should be scorporated in a different class (stub present)
source is strange since it doesn' need to be added to the scene but has to be passed at the tracer.
It would be possible to have several sources and simulate them indipendently BUT created photons need to hit device, so the relation light-source-device is crucial.

Long story short, probably it' better to keep the ligtsource in reactor class due to the relationship with the device.
Yet, all the experimental mathcing parameters and different lightsources should be initialized in a separated class for clarity and only the size/shape/orientation parameter should be initialized in reactor.

BOUND size increased to 20x20x20cm from (-100,-100,-100) to (100,100,100) for new unit of measure (1=mm)