Fix bounces
P.S. Now bounces are computed with absorption counter from state table using the last uuid of pid.
An alternative approach is to directly use the device propery to read bounces.
This alternative approach has the problem that, e.g. for channels photons the results of all the channels should be merged,
moreover, no distinction between solar and luminescent photons can be obtained.
Yet, for databases with several photons, operations on DB are unacceptably slow so a performance benchmarking of this
 solution is highly desirable. Eventually, a mixed implementation based on photons in db could be considered